build: Add --disable-Werror
We enable a bunch of compiler flags to trip common errors during
development. While this is very useful while hacking on Clutter, it
makes the life of people building Clutter on automated build systems
much harder; thus, we should have a configuration option to opt out of
the -Werror business.

GNOME has pretty much standardised on `--disable-Werror`, so we should
crib that configure option.